refractor subqueriesToDependentNode to generateNodeDependents
